changed BoardPos to short, remaned Constant to t_const, used std::array poiters to make classes lightweight, changed to rotationQ* in tetromino. that's it I'm done with performance improvement

+++ Classes/BoardPos.h
@@ -5,8 +5,8 @@
 
 struct BoardPos {
 	BoardPos() { x = 0; y = 0; }
-	BoardPos(int x, int y) { this->x = x; this->y = y; }
-	int x, y;
+	BoardPos(short x, short y) { this->x = x; this->y = y; }
+	short x, y;
 
 	BoardPos operator+ (const BoardPos& other)
 	{
